[#29] refactor: 멤버십 서버 전역에서 사용되는 클래스들의 패키지 재정의
