+++ membership/src/main/java/com/devcamp/flametalk/domain/profile/service/ProfileService.java
@@ -5,8 +5,8 @@ import com.devcamp.flametalk.domain.profile.domain.Profile;
 import com.devcamp.flametalk.domain.profile.domain.ProfileRepository;
 import com.devcamp.flametalk.domain.profile.dto.ProfileDetailResponse;
 import com.devcamp.flametalk.domain.profile.dto.ProfileRequest;
-import com.devcamp.flametalk.domain.profile.error.ErrorCode;
-import com.devcamp.flametalk.domain.profile.error.exception.EntityNotFoundException;
+import com.devcamp.flametalk.global.error.ErrorCode;
+import com.devcamp.flametalk.domain.profile.exception.EntityNotFoundException;
 import com.devcamp.flametalk.domain.user.domain.User;
 import com.devcamp.flametalk.domain.user.domain.UserRepository;
 import javax.transaction.Transactional;
