+++ membership/src/main/java/com/devcamp/flametalk/global/error/GlobalExceptionHandler.java
@@ -1,6 +1,6 @@
-package com.devcamp.flametalk.domain.profile.error;
+package com.devcamp.flametalk.global.error;
 
-import com.devcamp.flametalk.domain.profile.error.exception.EntityNotFoundException;
+import com.devcamp.flametalk.domain.profile.exception.EntityNotFoundException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.multipart.support.MissingServletRequestPartException;
 
 /**
- * 프로젝트 전역에 발생하는 Exception Handler 입니다.
+ * membership 서버 전역에 발생하는 Exception Handler 입니다.
  */
 @Slf4j
 @ControllerAdvice
